refactor: waiting queue domain model pojo객체로 변경, 로직 변경

+++ src/main/java/io/hhplus/concert/domain/common/ServicePolicy.java
@@ -5,5 +5,5 @@ public class ServicePolicy {
     public static final int WAITING_QUEUE_EXPIRED_MINUTES = 20;
     public static final int WAITING_QUEUE_ACTIVATE_COUNT = 1000;
 
-    public static final long WAITING_QUEUE_ACTIVATE_INTERVAL = 3_000;
+    public static final int WAITING_QUEUE_ACTIVATE_INTERVAL = 3_000;
 }
